refactor(query): place component tests with pdg
